[PA_3] Test Script implementaton and starter code
