+++ PA_3/solution/Makefile
@@ -1,0 +1,2 @@
+pq: pq.cpp
+	g++ pq.cpp -o pq
