+++ PA_3/solution/pq
